Move endpoints to Endpoints folder
