+++ Listem.API/Endpoints/ItemEndpoints.cs
@@ -1,10 +1,11 @@
 ﻿using Listem.API.Contracts;
+using Listem.API.Domain.Items;
 using Listem.API.Domain.Lists;
 using Listem.API.Middleware;
 using Microsoft.AspNetCore.Mvc;
 using static Listem.API.Utilities.EndpointUtilities;
 
-namespace Listem.API.Domain.Items;
+namespace Listem.API.Endpoints;
 
 [Route("api/items")]
 [ApiController]
