+++ Listem.API/Program.cs
@@ -2,6 +2,7 @@ using Listem.API.Contracts;
 using Listem.API.Domain.Categories;
 using Listem.API.Domain.Items;
 using Listem.API.Domain.Lists;
+using Listem.API.Endpoints;
 using Listem.API.Middleware;
 using Listem.API.Repositories;
 using Microsoft.AspNetCore.Identity;
